requirements.txt
chattts==0.1.1
edge_tts==6.1.12
Flask==3.0.3
Flask_SocketIO==5.3.7
funasr==1.1.6
gTTS==2.5.3
numpy==1.26.4
openai==1.45.0
PyAudio==0.2.14
pydub==0.25.1
PyYAML==6.0.2
silero_vad==5.1
torch==2.4.1
torchaudio==2.4.1
Flask-SocketIO~=5.3.7
requests~=2.32.3
schedule==1.2.2
beautifulsoup4==4.12.2
pygame==2.6.1
sounddevice==0.5.0
soundfile==0.12.1
playsound==1.3.0
langchain==0.3.2
langchain_chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.8
langchain-openai==0.2.1
langchain-text-splitters==0.3.0

# Web服务器
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Utils
pydantic>=2.0.0

